requirements.txt
openai
streamlit
python-dotenv
